+++ Quadrado.java
@@ -3,7 +3,7 @@ public class Quadrado extends FiguraGeometrica {
 	
 	public void calcularArea() {
 		
-		System.out.println("A area do quadrado e: " + this.getBase() * this.getBase());
+		System.out.println("A area do quadrado e: " + (long) (this.getBase() * this.getBase()));
 	}
 	
 	public Quadrado(double lado) {
