+++ Retangulo.java
@@ -3,13 +3,13 @@ public class Retangulo extends FiguraGeometrica{
 
 	public void calcularArea() {
 		
-		System.out.println("A area do retangulo e: " + this.getBase() * this.getAltura() ) ;
+		System.out.println("A area do retangulo e: " + (long) (this.getBase() * this.getAltura())) ;
 		
 	}
 	
 	public void perimetroRetangulo() {
 		
-		System.out.println("A area do retangulo e: " + (2 * this.getBase() ) + ( 2 * this.getAltura() ) ) ;
+		System.out.println("A area do retangulo e: " + (long) (2 * this.getBase() + 2 * this.getAltura()));
 		
 	}
